Serve static datasheets from data/
Some datasheets need to be served up locally as they don't exist online. Create
a directory that will serve up these static datasheets.

Note that this does not yet include a way to add datasheets to this dir, nor is
there a way to edit the URL field of a part to point it to something in the
directory.